requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
click==8.1.8
dill==0.4.0
discord.py==2.5.2
distro==1.9.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iso8601==2.1.0
isort==6.0.1
jiter==0.9.0
mccabe==0.7.0
multidict==6.4.3
mypy_extensions==1.1.0
openai==1.76.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pylint==3.3.6
python-dotenv==1.1.0
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.1
twitchio==2.10.0
typing-inspection==0.4.0
typing_extensions==4.13.2
unicode==2.9
Unidecode==1.4.0
yarl==1.20.0
